refactor: converted css into scss
